who know position control would be so hard and its not even in the motor controllers yet
this is a massive mess that needs to be cleaned up but I'm way to tired to spend another 30-40 min on this basically the correct abs encoder class is being used now, I really really hope that the fpga adc refresh rate is fast enough for this I think it is looking at the rio specs (it's 500 Ks/s and it's only reading 1 channel so....) anyways what this should do is switch all data gathering for the module angle onto the absolute encoder and also clean up the control flow for something fairly reasonable with the pid. I haven't decided how I want the normalization to go but I'll draw it up soon and get this fixed. abs encoder implementation should be done aside from finding the offset which is a Friday problem. woweee I wonder if anyone will read this :)