pid is still broken for module angle control but im not fixing it until the abs encoder stuff is resolved in person
I must have been pretty tired when I wrote the last commit because there are some very wrong things about that. The fpga is reading from FOUR channels, not one because there happen to be four modules, not one that we need data from. By my possibly flawed calculations, that makes the refresh rate 30hz ish which is slower than the rio (50hz) but hopefully can get the job done until I think of a better solution. The module angle pid is still a mess and will not work if the p is a non-zero value but until I know exactly what's happening with the units I'm not going to bother fixing it because I'll probably just have to change anything anyways.