big additions
reformatted some architecture, copied in the David auton stuff, finished the odometry updating stuff. The command termination needs to be added as well as being configured in robot container for it to actually be retrieved when running auton. Other sort of clean up stuff like that needs to be done so that it will run correctly.